Restructure commands.ts into commands dir
