+++ src/extension.ts
@@ -1,11 +1,9 @@
 "use strict";
 import * as vscode from "vscode";
-import * as commands from "./commands";
-import * as customfs from "./file_manipulation/file_system";
 import * as preferences from "./util/preferences";
 import * as compliance from "./util/compliance";
-import { robotType } from './templates/template_interpreter';
 import { displayChangelog } from './util/changelog';
+import { registerCommands } from "./commands/commands";
 
 var currentWorkspacePath: string;
 var currentWorkspaceFsPath: string;
@@ -19,86 +17,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
     // * Registering commands
     console.log("Registering commands");
-    let disposable = vscode.commands.registerCommand("kotlinforfrc.createNew", (file_path: any) => {
-        commands.createNew(file_path);
-    });
-
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand('kotlinforfrc.forceCompliance', async () => {
-        await commands.forceCompliance();
-    });
-
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand("kotlinforfrc.changeComplianceTestPref", () => {
-        commands.changeComplianceTestPref();
-    });
-
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand('kotlinforfrc.convertJavaProject', async () => {
-        console.log("Reading Robot.java");
-        // Check to make sure file paths are even there
-        try {
-            var robot_java: string = await customfs.readFile(currentWorkspaceFsPath + "/src/main/java/frc/robot/Robot.java");
-        }
-        catch (e) {
-            console.log(e);
-            vscode.window.showErrorMessage("Kotlin for FRC: Could not find Robot.java. You may have already converted this project or the correct directories are missing.");
-            return;
-        }
-        var current_robot_type: robotType = robotType.sample;
-
-        if (robot_java.includes("edu.wpi.first.wpilibj2.command.Command")) {
-            current_robot_type = robotType.command;
-            console.log("Command");
-        }
-        else if (robot_java.includes("edu.wpi.first.wpilibj.command.Command")) {
-            current_robot_type = robotType.old_command;
-            console.log("Old Command");
-        }
-        else if (robot_java.includes("edu.wpi.first.wpilibj.IterativeRobot")) {
-            current_robot_type = robotType.iterative;
-            console.log("Iterative");
-        }
-        else if (robot_java.includes("edu.wpi.first.wpilibj.SampleRobot")) {
-            current_robot_type = robotType.sample;
-            console.log("Sample");
-        }
-        else if (robot_java.includes("edu.wpi.first.wpilibj.TimedRobot")) {
-            if (robot_java.includes("edu.wpi.first.wpilibj.smartdashboard.SendableChooser")) {
-                current_robot_type = robotType.timed;
-                console.log("Timed");
-            }
-            else {
-                current_robot_type = robotType.timed_skeleton;
-                console.log("Timed skeleton");
-            }
-        }
-
-        commands.convertJavaProject(current_robot_type);
-    });
-    
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand("kotlinforfrc.showChangelog", () => {
-        commands.showChangelog();
-    });
-
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand("kotlinforfrc.toggleChangelog", () => {
-        commands.toggleChangelog(context);
-    });
-
-    context.subscriptions.push(disposable);
-
-    disposable = vscode.commands.registerCommand("kotlinforfrc.resetAutoShowChangelog", () => {
-        commands.resetAutoShowChangelog(context);
-    });
-
-    context.subscriptions.push(disposable);
+    await registerCommands(context);
 
     // * End registering commands
