+++ src/test/preferences.test.ts
